+++ R/internal.R
@@ -24,66 +24,6 @@ add_suffix <- function(x, y, sep = "_") {
   return(x)
 }
 
-#' Sorting files with time stamp and suffix
-#'
-#' When backups have been saved with a time stamp in the file's name and a
-#' suffix, in the case of more than one backup in one day.
-#'
-#' @param file A character value indicating the root of the backup's name. It
-#'     may include also the path relative to the working directory when the
-#'     backup is stored in a sub-folder.
-#' @param fext A character value indicating the file extension (including the
-#'     dot symbol).
-#'
-#' @return A data frame including the sorted names of backup files from the
-#' oldest to the newest.
-#'
-#' @keywords internal
-sort_backups <- function(file, f_timestamp = "%Y-%m-%d", fext = ".rda") {
-  path <- "."
-  if (grepl("/", file, fixed = TRUE)) {
-    path <- strsplit(file, "/", fixed = TRUE)[[1]]
-    file2 <- path[length(path)]
-    path <- paste(path[-length(path)], collapse = "/")
-  } else if (grepl("\\", file, fixed = TRUE)) {
-    path <- strsplit(file, "\\", fixed = TRUE)[[1]]
-    file2 <- path[length(path)]
-    path <- paste(path[-length(path)], collapse = "/")
-  } else {
-    file2 <- file
-  }
-  inFolder <- list.files(path = path, pattern = fext)
-  inFolder <- inFolder[grepl(file2, inFolder, fixed = TRUE)]
-  if (length(inFolder) == 0) {
-    stop("The requested backup is missing in the working directory.")
-  }
-  Name <- sub(fext, "", inFolder, fixed = TRUE)
-  Name <- strsplit(Name, "_", fixed = TRUE)
-  underscores <- gsub("_", "", file2)
-  underscores <- nchar(file2) - nchar(underscores)
-  Name <- lapply(Name, function(x, y) {
-    if (length(x) < (y + 3)) {
-      x <- c(x, rep_len("0", y - length(x) + 3))
-    }
-    return(x)
-  }, y = underscores)
-  OUT <- as.data.frame(do.call(rbind, Name), stringsAsFactors = FALSE)[
-    ,
-    c((underscores + 2):(underscores + 3))
-  ]
-  colnames(OUT) <- c("date", "suffix")
-  OUT$path <- path
-  OUT$filename <- inFolder
-  OUT$date <- as.Date(strptime(OUT$date, f_timestamp))
-  OUT <- OUT[!is.na(OUT$date), ]
-  OUT$suffix <- as.integer(OUT$suffix)
-  ## OUT <- OUT[with(OUT, order(date, suffix)), ]
-  OUT <- OUT[order(OUT$date, OUT$suffix), ]
-  OUT$order <- c(seq_len(nrow(OUT)))
-  return(OUT)
-}
-# TODO: add 'sep="_"' in the argument list
-
 #' Filling missed columns with NAs
 #'
 #' If columns of `y` are missed in `x`, the later gets these columns filled with