+++ tests/testthat/test-backup_object.R
@@ -1,11 +1,12 @@
 context("writing backups")
 
 test_that("Backups are written", {
-  backup_object(Easplist, file = "temp", stamp = FALSE, overwrite = TRUE)
-  expect_true(file.exists("temp.rda"))
-  backup_object(Easplist, file = "temp", overwrite = TRUE)
-  expect_true(file.exists(paste0("temp_", format(
-    Sys.Date(),
-    "%Y-%m-%d"
-  ), ".rda")))
+  back_log <- backup_object(Easplist, file = file.path(tempdir(), "temp"),
+      stamp = FALSE, overwrite = TRUE)
+  expect_true(file.exists(back_log["abspath"]))
+  back_log <- backup_object(Easplist, file = "temp", overwrite = TRUE)
+  expect_true(file.exists(back_log["abspath"]))
+  back_log <- backup_object(Easplist, file = "temp")
+  expect_true(file.exists(back_log["abspath"]))
+  expect_error(sort_backups(name = "apple", path = tempdir()))
 })